fix linux tests

+++ c/src/om_common.c
@@ -50,7 +50,7 @@ void om_common_copy_float_to_int32(uint64_t length, float scale_factor, float ad
             ((int32_t *)dst)[i] = INT32_MAX;
         } else {
             float scaled = val * scale_factor + add_offset;
-            ((int32_t *)dst)[i] = (int32_t)fmaxf(INT32_MIN, fminf(INT32_MAX, roundf(scaled)));
+            ((int32_t *)dst)[i] = (int32_t)fmaxf((float)INT32_MIN, fminf((float)INT32_MAX, roundf(scaled)));
         }
     }
 }
@@ -62,7 +62,7 @@ void om_common_copy_double_to_int64(uint64_t length, float scale_factor, float a
             ((int64_t *)dst)[i] = INT64_MAX;
         } else {
             double scaled = val * (double)scale_factor + (double)add_offset;
-            ((int64_t *)dst)[i] = (int64_t)fmax(INT64_MIN, fmin(INT64_MAX, round(scaled)));
+            ((int64_t *)dst)[i] = (int64_t)fmax((float)INT64_MIN, fmin((float)INT64_MAX, round(scaled)));
         }
     }
 }
@@ -74,7 +74,7 @@ void om_common_copy_float_to_int16_log10(uint64_t length, float scale_factor, fl
             ((int16_t *)dst)[i] = INT16_MAX;
         } else {
             float scaled = log10f(1 + val) * scale_factor;
-            ((int16_t *)dst)[i] = (int16_t)fmaxf(INT16_MIN, fminf(INT16_MAX, roundf(scaled)));
+            ((int16_t *)dst)[i] = (int16_t)fmaxf((float)INT16_MIN, fminf((float)INT16_MAX, roundf(scaled)));
         }
     }
 }
